requirements.txt
flask
flask-cors
Pillow
torch
torchvision
transformers
sentence-transformers
numpy
requests
python-dotenv
gunicorn
werkzeug